Rework failure reporting architecture
The CIQ simulator appears to have issues when exceptions cross a
Lang.Method boundary. Instead throw and capture exceptions within the
Test call stack and return them in an array. For the test runner to use.

Still using exceptions because call stacks can be useful to determine
which expectations failed.