Start working on Mock implementation of failures
The change in test architecture away from wildly throwing exceptions,
requires some architecture changes in the mocking frame work. Move
common "Failure List" logic out of test and into a new class, because
mocks will need to implement a similar throw and catch type of handling.

Also add an exception in TestMain if any tests fail. This ensures that
the simulator exits with a non-zero code and fails on travis.